[ICACHE] Fixed timing issues
Prefetcher was sometimes triggering a double cache miss as the request for decoding instructions was timed. Prefetcher now differentiates timed and untimed fetches, and cache and memory models are ignoring untimed requests.

Added a fake prefetch of the next instruction during a jump or branch to model the fact that the core will always fetch the next instruction which can trigger a cache miss in some cases